+++ src/narf/gl/context.cpp
@@ -30,7 +30,7 @@ void narf::gl::Context::setVsync(bool enabled) {
 }
 
 
-bool narf::gl::Context::setDisplayMode(const char *title, uint32_t width, uint32_t height, bool fullscreen)
+bool narf::gl::Context::setDisplayMode(const char *title, int32_t width, int32_t height, bool fullscreen)
 {
 	Uint32 flags = SDL_WINDOW_OPENGL | SDL_WINDOW_RESIZABLE;
 	if (fullscreen) {
@@ -89,17 +89,17 @@ void narf::gl::Context::updateViewport() {
 }
 
 
-uint32_t narf::gl::Context::width() const {
+int32_t narf::gl::Context::width() const {
 	int w, h;
 	SDL_GetWindowSize(window_, &w, &h);
-	return static_cast<uint32_t>(w);
+	return w;
 }
 
 
-uint32_t narf::gl::Context::height() const {
+int32_t narf::gl::Context::height() const {
 	int w, h;
 	SDL_GetWindowSize(window_, &w, &h);
-	return static_cast<uint32_t>(h);
+	return h;
 }
